requirements.txt
aiogram==3.20.0
python-docx==1.1.2
langchain_gigachat==0.3.10
pandas==2.2.3